Refactor Profile module and add support for app shortcuts
Now the a Profile also supports a set of application shortcuts.
Previously they were global only. Additionally this refactoring
introduces a clear vocabulary for Profiles.
The active profile is currently active one. The designated
profile is the profile which is not necessary active but should
according to the screen configuration.